pyproject.toml
[tool.poetry]
name = "bababos"
version = "0.1.0"
description = "Bababos pricing mechanism"
authors = ["Fathur Rohman <hi.fathur.rohman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
django-mptt = "^0.16.0"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
django-extensions = "^3.2.3"
ipython = "^8.22.1"
djangorestframework = "^3.14.0"
django-filter = "^23.5"
django-softdelete = "^0.10.5"
pandas = "^2.2.1"
parameterized = "^0.9.0"
tabulate = "^0.9.0"
babel = "^2.14.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
flake8 = "^7.0.0"
pre-commit = "^3.6.2"
faker = "^23.3.0"
pytest = "^8.0.2"
pytest-django = "^4.8.0"
pytest-env = "^1.1.3"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bababos.settings"
python_files = ["test_*.py", "*_test.py"]
filterwarnings = "ignore::DeprecationWarning"
env = [
    "SECRET_KEY=fasd8n7fnnq34",
    "DB_DRIVER=postgresql",
    "DB_HOST=localhost",
    "DB_USER=fathur",
    "DB_PASSWORD=plokijuh",
    "DB_NAME=bababos_test",
    "DB_PORT=5432",
]
